+++ t0016Next/myapp/src/components/form/CreateContentForm/ErrorMessage.tsx
@@ -1,0 +1,9 @@
+import { FieldError, useForm } from "react-hook-form"
+
+type ErrorMessageProps = {
+  errorField: FieldError | undefined
+}
+
+export const ErrorMessage = ({ errorField }: ErrorMessageProps) => {
+  return <span className="text-red-500">{errorField?.message}</span>
+}
